third_party/blink/web_tests/external/wpt/svg/text/scripted/getrotationofchar-expected.txt

This is a testharness.js-based test.
[FAIL] 'text-orientation: mixed' - sideways-rl
  assert_equals: 月 expected 90 but got 0
[FAIL] 'text-orientation: upright' - sideways-rl
  assert_equals: T expected 90 but got 0
[FAIL] Rotation attribute - sideways-rl
  assert_equals: 木 expected 270.5 but got 180.5
[FAIL] <textPath> - sideways-rl
  assert_equals: 金 expected 45 but got 315
[FAIL] 'text-orientation: mixed' - sideways-lr
  assert_equals: M expected -90 but got 90
[FAIL] 'text-orientation: upright' - sideways-lr
  assert_equals: T expected -90 but got 0
[FAIL] 'text-orientation: sideways' - sideways-lr
  assert_equals: W expected -90 but got 90
[FAIL] Rotation attribute - sideways-lr
  assert_equals: t expected 270 but got 90
[FAIL] <textPath> - sideways-lr
  assert_equals: F expected 225 but got 45
Harness: the test ran to completion.
